Duty: Added grenades & changed damage to floats

+++ Demo-Disc-Vol-1/Assets/Scripts/Duty/PlayerControllerDuty.cs
@@ -214,7 +214,7 @@ public class PlayerControllerDuty : MonoBehaviour
         displayAmmo();
     }
 
-    public void HandleDamage(int damage)
+    public void HandleDamage(float damage)
     {
         health -= damage;
         healthFill.fillAmount = health / maxHealth;
